requirements.txt
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
matplotlib==3.9.0
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.5.0
scikit-image==0.22.0
scipy==1.13.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
gdown==5.2.0